+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Fixed
 
 * Remove duplicated code
 * Properly handle non-ASCII short file names / 8DOT3
+* Mark dir/file entries as empty on deletion
 
 Changed
 ~~~~~~~